requirements.txt
python-telegram-bot
asyncpg
aiohttp
Flask
uvicorn
